chore(`getaddressdeltas`): move things around
